Simplified Main class, separation into DemoApplication
To me, it seems odd to have non-static contents of the Main class, and to create an instance of Main. Therefore I have made a super-simple Main, and put the logic into a separate class, DemoApplication.

I have also made a few simplifiations of the demo application.

The old Main is now OldMain.
